make newconnetor an example
